# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SynergyVAHouse API Documentation",
    "version": "1.0.0",
    "description": "# SynergyVAHouse API\n\nA comprehensive task management and workflow automation platform with webhook integrations, AI assistance, and RBAC.\n\n## Authentication\n\nThe API supports multiple authentication methods:\n\n### 1. JWT Bearer Tokens\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n### 2. API Keys\n```\nX-API-Key: <your_api_key>\n```\n\n### 3. Webhook Signatures\nFor webhook endpoints (Conduit and SuiteOp):\n```\nX-Conduit-Signature: sha256=<hmac_signature>\nX-SuiteOp-Signature: sha256=<hmac_signature>\n```\n\n### 4. Slack Signatures\nFor Slack events:\n```\nX-Slack-Signature: v0=<hmac_signature>\nX-Slack-Request-Timestamp: <unix_timestamp>\n```\n\n## Rate Limiting\n\nAPI endpoints are rate-limited to ensure fair usage. Please implement exponential backoff for retries.\n\n## Idempotency\n\nWebhook endpoints support idempotency through event IDs. Duplicate events with the same ID will return a 200 status with `status: 'duplicate'`.",
    "contact": {
      "name": "SynergyVAHouse Team"
    },
    "license": {
      "name": "Private"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "API Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Conduit-Signature",
        "description": "HMAC-SHA256 signature for webhook verification"
      },
      "slackSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Slack-Signature",
        "description": "Slack request signature verification"
      }
    },
    "schemas": {
...
      "AuthTokens": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token (15min expiry)"
          },
          "refreshToken": {
            "type": "string",
            "description": "Refresh token (7d expiry)"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
...
    },
...
  },
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "User login",
        "description": "Authenticates a user with email and password credentials.\nReturns JWT access token (15min expiry) and refresh token (7d expiry).\n\nNote: Current implementation is simplified for demo purposes.\nIn production, password validation should be implemented.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "User email address",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "User password",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokens"
                }
              }
            }
          },
          "400": {
            "description": "Email is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Email is required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid credentials"
                }
              }
            }
          },
          "403": {
            "description": "Account is disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Account is disabled"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Login failed"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh access token",
        "description": "Exchanges a valid refresh token for a new access token and refresh token.\nImplements automatic token rotation for enhanced security.\nThe old refresh token is revoked after use.\n",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "refreshToken"
                ],
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "Valid refresh token",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokens"
                }
              }
            }
          },
          "400": {
            "description": "Refresh token is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Refresh token is required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired refresh token"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout from current session",
        "description": "Revokes the provided refresh token, logging out the user from the current session.\nThe user will remain logged in on other devices.\n",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string",
                    "description": "Refresh token to revoke"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out successfully"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logout from all devices",
        "description": "Revokes all refresh tokens for the authenticated user, logging them out from all devices.\nRequires authentication via Bearer token or API key.\n",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out from all devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out from all devices"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/api/auth/slack/authorize": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Initiate Slack OAuth flow",
        "description": "Redirects to Slack OAuth authorization URL",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            },
            "description": "Optional state parameter for OAuth flow"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to Slack OAuth URL"
          },
          "500": {
            "description": "Slack OAuth not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/slack/callback": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Slack OAuth callback",
        "description": "Handles the OAuth callback from Slack.\nExchanges authorization code for user info and creates/updates user account.\nRedirects to frontend with JWT tokens.\n",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Authorization code from Slack"
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            },
            "description": "State parameter from authorization request"
          },
          {
            "in": "query",
            "name": "error",
            "schema": {
              "type": "string"
            },
            "description": "Error code if authorization failed"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to frontend with tokens"
          },
          "400": {
            "description": "Authorization code is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/user": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get current user",
        "description": "Returns the currently authenticated user's information",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "authType": {
                          "type": "string",
                          "description": "Authentication method used",
                          "enum": [
                            "jwt",
                            "apiKey",
                            "replit",
                            "slack",
                            "dev"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
  }
}